Refactor: 예외 처리 리펙토링

+++ src/main/java/com/capstone/rentit/item/service/ItemService.java
@@ -1,8 +1,12 @@
 package com.capstone.rentit.item.service;
 
+import com.capstone.rentit.common.CommonResponse;
 import com.capstone.rentit.item.domain.Item;
 import com.capstone.rentit.item.dto.*;
+import com.capstone.rentit.item.exception.ItemNotFoundException;
+import com.capstone.rentit.item.exception.ItemUnauthorizedException;
 import com.capstone.rentit.item.repository.ItemRepository;
+import com.capstone.rentit.member.dto.MemberDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -47,19 +51,35 @@ public class ItemService {
 
     @Transactional(readOnly = true)
     public ItemDto getItem(Long itemId) {
-        Item item = itemRepository.findById(itemId)
-                .orElseThrow(() -> new RuntimeException("Item not found"));
+        Item item = findItem(itemId);
         return ItemDto.fromEntity(item);
     }
 
-    public void updateItem(Long itemId, ItemUpdateForm form) {
-        Item item = itemRepository.findById(itemId)
-                .orElseThrow(() -> new RuntimeException("Item not found"));
+    public void updateItem(MemberDto loginMember, Long itemId, ItemUpdateForm form) {
+        Item item = findItem(itemId);
+        assertOwner(item, loginMember.getId());
+
         item.updateItem(form);
         itemRepository.save(item);
     }
 
-    public void deleteItem(Long itemId) {
+    public void deleteItem(MemberDto loginMember, Long itemId) {
+        Item item = findItem(itemId);
+        assertOwner(item, loginMember.getId());
+
         itemRepository.deleteById(itemId);
+    }
+
+    private Item findItem(Long itemId) {
+        return itemRepository.findById(itemId)
+                .orElseThrow(() ->
+                        new ItemNotFoundException("존재하지 않는 물품입니다.")
+                );
+    }
+
+    private void assertOwner(Item item, Long userId) {
+        if (!item.getOwnerId().equals(userId)) {
+            throw new ItemUnauthorizedException("자신의 소유 물품이 아닙니다.");
+        }
     }
 }
